refactoring getVersionFromButton: more dynamic
